pyproject.toml
[project]
name = "lib-bi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "cachetools>=5.4.0",
    "certifi>=2024.7.4",
    "charset-normalizer>=3.3.2",
    "google-api-core>=2.19.1",
    "google-auth>=2.33.0",
    "google-auth-oauthlib>=1.2.1",
    "google-cloud-bigquery>=3.26.0",
    "google-cloud-bigquery-storage>=2.26.0",
    "google-cloud-storage>=2.18.2",
    "google-cloud-core>=2.4.1",
    "google-crc32c>=1.5.0",
    "google-resumable-media>=2.7.2",
    "googleapis-common-protos>=1.63.2",
    "grpcio>=1.65.4",
    "grpcio-status>=1.65.4",
    "gspread>=6.1.2",
    "idna>=3.7",
    "more-itertools>=10.2.0",
    "numpy>=1.26.4",
    "oauthlib>=3.2.2",
    "oauth2client>=4.1.3",
    "packaging>=24.1",
    "polars>=1.4.1",
    "proto-plus>=1.24.0",
    "protobuf>=5.27.3",
    "pyarrow>=17.0.0",
    "pyasn1>=0.6.0",
    "pyasn1_modules>=0.4.0",
    "pydantic>=2.6.1",
    "pydantic_core>=2.16.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pytz>=2024.1",
    "requests>=2.32.3",
    "requests-oauthlib>=2.0.0",
    "rsa>=4.9",
    "six>=1.16.0",
    "typing_extensions>=4.12.2",
    "urllib3>=2.2.2",
    "xlsx2csv>=0.8.3",
    "openpyxl>=3.0.10",
    "pandas>=2.2.3",
    "rich>=13.8.1",
    "pysftp>=0.2.9",
    "paramiko>=3.5.0",
    "cryptography>=43.0.3",
    "db-dtypes>=1.3.0",
    "google-api-python-client>=2.153.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
